+++ protostar/commands/test/test_command.py
@@ -8,7 +8,7 @@ from protostar.commands.test.test_collector_summary_formatter import (
     format_test_collector_summary,
 )
 from protostar.commands.test.test_result_formatter import format_test_result
-from protostar.commands.test.test_results import TestResult
+from protostar.test_runner.test_results import TestResult
 from protostar.commands.test.test_runner import TestRunner
 from protostar.commands.test.test_scheduler import TestScheduler
 from protostar.commands.test.testing_live_logger import TestingLiveLogger
