+++ protostar/commands/test/test_result_formatter.py
@@ -4,20 +4,19 @@ from typing import Callable, Dict, List
 from protostar.commands.test.test_environment_exceptions import ExceptionMetadata
 from protostar.commands.test.test_output_recorder import OutputName, format_output_name
 from protostar.protostar_exception import UNEXPECTED_PROTOSTAR_ERROR_MSG
-from protostar.utils.log_color_provider import log_color_provider
-
-from .test_results import (
+from protostar.test_runner import (
     BrokenFuzzTestCaseResult,
+    BrokenTestCaseResult,
     BrokenTestSuiteResult,
     FailedFuzzTestCaseResult,
     FailedTestCaseResult,
-    BrokenTestCaseResult,
     PassedFuzzTestCaseResult,
     PassedTestCaseResult,
     SkippedTestCaseResult,
     TestResult,
     UnexpectedBrokenTestSuiteResult,
 )
+from protostar.utils.log_color_provider import log_color_provider
 
 LogCallback = Callable[[str], None]
