+++ protostar/test_runner/__init__.py
@@ -1,1 +1,20 @@
 from .test_collector import TestCollector
+from .test_results import (
+    BrokenFuzzTestCaseResult,
+    BrokenSetupCaseResult,
+    BrokenTestCaseResult,
+    BrokenTestSuiteResult,
+    FailedFuzzTestCaseResult,
+    FailedTestCaseResult,
+    PassedFuzzTestCaseResult,
+    PassedSetupCaseResult,
+    PassedTestCaseResult,
+    SetupCaseResult,
+    SkippedSetupCaseResult,
+    SkippedTestCaseResult,
+    TestCaseResult,
+    TestResult,
+    TimedTestCaseResult,
+    TimedTestResult,
+    UnexpectedBrokenTestSuiteResult,
+)
