+++ tests/integration/cheatcodes/skip/skip_test.py
@@ -4,7 +4,7 @@ from typing import List
 import pytest
 
 from protostar.commands.test.test_result_formatter import format_test_result
-from protostar.commands.test.test_results import SkippedTestCaseResult
+from protostar.test_runner.test_results import SkippedTestCaseResult
 
 from tests.integration.conftest import (
     RunCairoTestRunnerFixture,
