+++ tests/integration/fuzzing_strategies/fuzzing_strategies_test.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import pytest
 
-from protostar.commands.test.test_results import PassedFuzzTestCaseResult
+from protostar.test_runner.test_results import PassedFuzzTestCaseResult
 from tests.integration.conftest import (
     RunCairoTestRunnerFixture,
     assert_cairo_test_cases,
